De-coupled the Notification Bar update and Drink Notification sound clip code from PowerHourService. Used BroadcastManager to catch the ProgressUpdate and MusicUpdate events. Added some gitignore for IntelliJ. Used IntelliJ analysis to DRY up code

+++ src/gilday/android/powerhour/data/InitializePlaylistTask.java
@@ -205,7 +205,7 @@ public abstract class InitializePlaylistTask extends AsyncTask<Void, Void, Integ
 		
 		public QuickLoadImportCursorIterator() {
     		int importSize = importCursor.getCount();
-    		bucketSize = (int) (importSize / QUICKLOAD_THRESHOLD);
+    		bucketSize = importSize / QUICKLOAD_THRESHOLD;
     		rand = new Random();
 		}
